+++ apps/web/package.json
@@ -11,17 +11,12 @@
   "dependencies": {
     "@radix-ui/colors": "1.0.1",
     "@radix-ui/react-slot": "1.0.2",
-    "@sindresorhus/is": "5.5.2",
-    "@supabase/supabase-js": "2.26.0",
     "@vercel/analytics": "1.0.1",
     "classnames": "2.3.2",
-    "date-fns": "^2.29.3",
     "next": "13.5.3",
-    "prism-react-renderer": "1.3.5",
+    "prism-react-renderer": "2.1.0",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "resend": "1.1.0",
-    "zod": "3.22.2"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "13.5.3",
