--- apps/web/src/components/anchor/index.ts
@@ -1,1 +1,0 @@
-export * from "./anchor";
