--- apps/web/src/components/badge/index.ts
@@ -1,1 +1,0 @@
-export * from "./badge";
