--- apps/web/src/components/button/index.ts
@@ -1,1 +1,0 @@
-export * from "./button";
