+++ apps/web/src/components/heading.tsx
@@ -1,33 +1,43 @@
 import * as SlotPrimitive from "@radix-ui/react-slot";
 import classnames from "classnames";
 import * as React from "react";
-import type { As } from "../../utils";
-import { unreachable } from "../../utils";
+import type { As } from "../utils/as";
+import { unreachable } from "../utils/unreachable";
 
-export type TextSize = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9";
-export type TextColor = "gray" | "white";
-export type TextTransform = "uppercase" | "lowercase" | "capitalize";
-export type TextWeight = "normal" | "medium";
+export type HeadingSize =
+  | "1"
+  | "2"
+  | "3"
+  | "4"
+  | "5"
+  | "6"
+  | "7"
+  | "8"
+  | "9"
+  | "10";
+export type HeadingColor = "white" | "gray";
+export type HeadingWeight = "medium" | "bold";
 
-interface TextOwnProps {
-  size?: TextSize;
-  color?: TextColor;
-  transform?: TextTransform;
-  weight?: TextWeight;
+interface HeadingOwnProps {
+  size?: HeadingSize;
+  color?: HeadingColor;
+  weight?: HeadingWeight;
 }
 
-type TextProps = As<"span", "div", "p"> & TextOwnProps;
+type HeadingProps = As<"h1", "h2", "h3", "h4", "h5", "h6"> & HeadingOwnProps;
 
-export const Text = React.forwardRef<HTMLSpanElement, Readonly<TextProps>>(
+export const Heading = React.forwardRef<
+  HTMLHeadingElement,
+  Readonly<HeadingProps>
+>(
   (
     {
-      as: Tag = "span",
-      size = "2",
-      color = "gray",
-      transform,
-      weight = "normal",
+      as: Tag = "h1",
+      size = "3",
       className,
+      color = "white",
       children,
+      weight = "bold",
       ...props
     },
     forwardedRef,
@@ -35,7 +45,6 @@ export const Text = React.forwardRef<HTMLSpanElement, Readonly<TextProps>>(
     <SlotPrimitive.Slot
       className={classnames(
         className,
-        transform,
         getSizesClassNames(size),
         getColorClassNames(color),
         getWeightClassNames(weight),
@@ -48,19 +57,19 @@ export const Text = React.forwardRef<HTMLSpanElement, Readonly<TextProps>>(
   ),
 );
 
-const getSizesClassNames = (size: TextSize | undefined) => {
+const getSizesClassNames = (size: HeadingSize | undefined) => {
   switch (size) {
     case "1":
       return "text-xs";
-    case undefined:
     case "2":
       return "text-sm";
+    case undefined:
     case "3":
       return "text-base";
     case "4":
       return "text-lg";
     case "5":
-      return ["text-17px", "md:text-xl tracking-[-0.16px]"];
+      return "text-xl tracking-[-0.16px]";
     case "6":
       return "text-2xl tracking-[-0.288px]";
     case "7":
@@ -69,33 +78,38 @@ const getSizesClassNames = (size: TextSize | undefined) => {
       return "text-[35px] leading-[42px] tracking-[-0.64px]";
     case "9":
       return "text-6xl leading-[73px] tracking-[-0.896px]";
+    case "10":
+      return [
+        "text-[38px] leading-[46px]",
+        "md:text-[70px] md:leading-[85px] tracking-[-1.024px;]",
+      ];
     default:
       return unreachable(size);
   }
 };
 
-const getColorClassNames = (color: TextColor | undefined) => {
+const getColorClassNames = (color: HeadingColor | undefined) => {
   switch (color) {
-    case "white":
-      return "text-slate-12";
-    case undefined:
     case "gray":
       return "text-slate-11";
+    case "white":
+    case undefined:
+      return "text-slate-12";
     default:
       return unreachable(color);
   }
 };
 
-const getWeightClassNames = (weight: TextWeight | undefined) => {
+const getWeightClassNames = (weight: HeadingWeight | undefined) => {
   switch (weight) {
-    case undefined:
-    case "normal":
-      return "font-normal";
     case "medium":
       return "font-medium";
+    case "bold":
+    case undefined:
+      return "font-bold";
     default:
       return unreachable(weight);
   }
 };
 
-Text.displayName = "Text";
+Heading.displayName = "Heading";
