--- apps/web/src/components/heading/index.ts
@@ -1,1 +1,0 @@
-export * from "./heading";
