--- apps/web/src/components/icon-button/index.ts
@@ -1,1 +1,0 @@
-export * from "./icon-button";
