--- apps/web/src/components/logo/index.ts
@@ -1,1 +1,0 @@
-export * from "./logo";
