+++ apps/web/src/components/text.tsx
@@ -1,43 +1,33 @@
 import * as SlotPrimitive from "@radix-ui/react-slot";
 import classnames from "classnames";
 import * as React from "react";
-import type { As } from "../../utils";
-import { unreachable } from "../../utils";
+import type { As } from "../utils/as";
+import { unreachable } from "../utils/unreachable";
 
-export type HeadingSize =
-  | "1"
-  | "2"
-  | "3"
-  | "4"
-  | "5"
-  | "6"
-  | "7"
-  | "8"
-  | "9"
-  | "10";
-export type HeadingColor = "white" | "gray";
-export type HeadingWeight = "medium" | "bold";
+export type TextSize = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9";
+export type TextColor = "gray" | "white";
+export type TextTransform = "uppercase" | "lowercase" | "capitalize";
+export type TextWeight = "normal" | "medium";
 
-interface HeadingOwnProps {
-  size?: HeadingSize;
-  color?: HeadingColor;
-  weight?: HeadingWeight;
+interface TextOwnProps {
+  size?: TextSize;
+  color?: TextColor;
+  transform?: TextTransform;
+  weight?: TextWeight;
 }
 
-type HeadingProps = As<"h1", "h2", "h3", "h4", "h5", "h6"> & HeadingOwnProps;
+type TextProps = As<"span", "div", "p"> & TextOwnProps;
 
-export const Heading = React.forwardRef<
-  HTMLHeadingElement,
-  Readonly<HeadingProps>
->(
+export const Text = React.forwardRef<HTMLSpanElement, Readonly<TextProps>>(
   (
     {
-      as: Tag = "h1",
-      size = "3",
+      as: Tag = "span",
+      size = "2",
+      color = "gray",
+      transform,
+      weight = "normal",
       className,
-      color = "white",
       children,
-      weight = "bold",
       ...props
     },
     forwardedRef,
@@ -45,6 +35,7 @@ export const Heading = React.forwardRef<
     <SlotPrimitive.Slot
       className={classnames(
         className,
+        transform,
         getSizesClassNames(size),
         getColorClassNames(color),
         getWeightClassNames(weight),
@@ -57,19 +48,19 @@ export const Heading = React.forwardRef<
   ),
 );
 
-const getSizesClassNames = (size: HeadingSize | undefined) => {
+const getSizesClassNames = (size: TextSize | undefined) => {
   switch (size) {
     case "1":
       return "text-xs";
+    case undefined:
     case "2":
       return "text-sm";
-    case undefined:
     case "3":
       return "text-base";
     case "4":
       return "text-lg";
     case "5":
-      return "text-xl tracking-[-0.16px]";
+      return ["text-17px", "md:text-xl tracking-[-0.16px]"];
     case "6":
       return "text-2xl tracking-[-0.288px]";
     case "7":
@@ -78,38 +69,33 @@ const getSizesClassNames = (size: HeadingSize | undefined) => {
       return "text-[35px] leading-[42px] tracking-[-0.64px]";
     case "9":
       return "text-6xl leading-[73px] tracking-[-0.896px]";
-    case "10":
-      return [
-        "text-[38px] leading-[46px]",
-        "md:text-[70px] md:leading-[85px] tracking-[-1.024px;]",
-      ];
     default:
       return unreachable(size);
   }
 };
 
-const getColorClassNames = (color: HeadingColor | undefined) => {
+const getColorClassNames = (color: TextColor | undefined) => {
   switch (color) {
+    case "white":
+      return "text-slate-12";
+    case undefined:
     case "gray":
       return "text-slate-11";
-    case "white":
-    case undefined:
-      return "text-slate-12";
     default:
       return unreachable(color);
   }
 };
 
-const getWeightClassNames = (weight: HeadingWeight | undefined) => {
+const getWeightClassNames = (weight: TextWeight | undefined) => {
   switch (weight) {
+    case undefined:
+    case "normal":
+      return "font-normal";
     case "medium":
       return "font-medium";
-    case "bold":
-    case undefined:
-      return "font-bold";
     default:
       return unreachable(weight);
   }
 };
 
-Heading.displayName = "Heading";
+Text.displayName = "Text";
