--- apps/web/src/components/text/index.ts
@@ -1,1 +1,0 @@
-export * from "./text";
