+++ apps/web/src/pages/_document.tsx
@@ -1,27 +1,18 @@
-import type { DocumentContext, DocumentInitialProps } from "next/document";
-import Document, { Head, Html, Main, NextScript } from "next/document";
+import { Html, Head, Main, NextScript } from "next/document";
 
-class MyDocument extends Document {
-  static async getInitialProps(
-    ctx: DocumentContext,
-  ): Promise<DocumentInitialProps> {
-    const initialProps = await Document.getInitialProps(ctx);
-
-    return initialProps;
-  }
-
-  render() {
-    return (
-      <Html className="text-slate-12 selection:bg-cyan-5 selection:text-cyan-12 bg-black">
-        <Head />
-
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
+const MyDocument = () => {
+  return (
+    <Html
+      className="text-slate-12 selection:bg-cyan-5 selection:text-cyan-12 bg-black"
+      lang="en"
+    >
+      <Head />
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+};
 
 export default MyDocument;
