--- apps/web/src/pages/api/send/test.ts
@@ -1,19 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-import { Resend } from "resend";
-
-const resend = new Resend("re_g4UAGCCP_PcWALk33riV1K9TGkJRob9W2");
-// const supabase = createClient(supabaseUrl, supabaseKey);
-
-export default async function sendTest(
-  _req: NextApiRequest,
-  res: NextApiResponse,
-) {
-  await resend.emails.create({
-    from: "React Email <preview@react.email>",
-    to: ["test-7juuvde18@srv1.mail-tester.com"],
-    subject: "Tarifário Atualizado | VOA Hotel Rezende",
-    html: `<!DOCTYPE htmlPUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd"><style>@import url(&#x27;https://fonts.googleapis.com/css2?family=Roboto&amp;display=swap&#x27;); body{font-family: &#x27;Roboto&#x27;, sans-serif; background: #d1d5db; font-size: 14px,}table, td{border-collapse: collapse; mso-table-lspace: 0pt; mso-table-rspace: 0pt;}img{border: 0; height: auto; line-height: 100%; outline: none; text-decoration: none; -ms-interpolation-mode: bicubic;}p{display: block; margin: 0;}</style><div style="width:700px;margin:0 auto;padding:0;text-align:center"><div style="margin-top:10px;margin-bottom:25px;background:#FFFFFF;padding-top:20px;padding-bottom:40px"><h1 style="color:#252f3f;font-size:22px;line-height:25px;margin-left:15px;margin-right:15px">Alteração de tarifário do <br/>VOA Hotel Rezende</h1><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Luxo Triplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 320 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 337</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 368 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 387</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 420 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 442</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 291 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 306</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 382 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 402</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 320 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 337</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 368 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 387</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 420 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 442</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 332 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 326</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 382 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 375</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 428</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 330 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 291</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 379 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 334</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 433 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 382</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Luxo Casal</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 320 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 337</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 368 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 387</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 291 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 306</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 320 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 337</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 368 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 387</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 332 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 326</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 382 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 375</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 330 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 291</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 379 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 334</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Superior Quíntuplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th><th>PAX 4</th><th>PAX 5</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 456 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 480</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 456 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 480</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 456 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 480</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 456 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 480</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 456 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 480</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 502 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 528</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 522 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 511</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 522 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 511</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 522 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 511</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 522 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 511</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 522 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 511</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 517 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 456</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 517 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 456</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 517 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 456</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 517 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 456</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 517 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 456</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Superior Quádruplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th><th>PAX 4</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 437 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 460</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 454 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 445</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 454 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 445</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 454 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 445</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 454 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 445</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 451 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 397</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 451 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 397</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 451 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 397</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 451 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 397</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Superior Triplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 366 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 385</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 247 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 260</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 304 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 320</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 333 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 350</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 366 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 385</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 282 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 277</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 348 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 340</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 380 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 372</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 247</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 345 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 304</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 377 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 333</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Superior Casal</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 247 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 260</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 304 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 320</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 334 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 352</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 282 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 277</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 348 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 340</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 247</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 345 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 304</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Standard Quádruplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th><th>PAX 4</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 380 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 400</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 380 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 400</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 380 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 400</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 380 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 400</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 418 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 440</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 435 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 426</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 435 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 426</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 435 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 426</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 435 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 426</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 431 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 431 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 431 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 431 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Standard Triplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 234 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 246</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 270 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 284</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 316 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 332</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 213 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 224</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 245 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 258</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 287 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 302</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 234 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 246</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 270 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 284</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 316 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 332</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 243 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 238</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 275</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 328 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 321</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 241 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 213</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 278 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 245</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 326 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 287</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Standard Casal</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 234 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 246</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 270 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 284</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 213 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 224</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 245 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 258</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 234 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 246</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 270 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 284</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 243 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 238</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 275</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 241 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 213</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 278 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 245</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Econômico Quadrúplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th><th>PAX 4</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 361 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 361 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 361 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 361 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 380</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 397 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 418</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 413 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 404</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 413 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 404</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 413 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 404</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 413 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 404</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 410 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 361</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 410 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 361</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 410 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 361</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 410 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 361</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Econômico Triplo</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th><th>PAX 3</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 209 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 220</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 359 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 378</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 190 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 200</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 247 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 260</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 327 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 344</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 209 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 220</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 359 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 378</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 217 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 213</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 282 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 277</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 374 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 366</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 216 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 190</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 247</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 371 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 327</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div><div style="margin:28px 0"><div style="border-bottom:1px solid rgb(231, 231, 231);padding:18px"><div style="text-align:left;margin-bottom:8px"><h2 style="color:#252f3f;font-weight:bold;text-transform:uppercase;font-size:18px;margin-bottom:4px">Econômico Casal</h2><p>01/02/2023 - 23/02/2023</p></div><table style="width:100%;table-layout:fixed;border:1px solid rgb(231, 231, 231);border-collapse:collapse;margin:0px;padding:0px"><thead><tr style="background:rgb(250, 250, 250);border-bottom:1px solid rgb(231, 231, 231)"><th style="width:30%;text-align:left;padding:12px">DATA</th><th>PAX 1</th><th>PAX 2</th></tr></thead><tbody><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">01/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 209 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 220</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">02/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 190 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 200</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 247 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 260</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">07/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 209 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 220</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/down.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 272 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 286</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">22/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 217 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 213</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 282 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 277</span></span></td></tr></tbody></table></td></tr><tr style="border-bottom:1px solid rgb(231, 231, 231);color:#6b7280;font-size:16px;line-height:14px"><td style="width:30%;text-align:left;padding:12px">23/02/2023</td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 216 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 190</span></span></td></tr></tbody></table></td><td style="padding:14px;text-align:center"><table style="margin:0 auto"><tbody><tr><td><img src="https://rms.voahoteis.io/up.png" alt="Icon"/></td><td><span style="text-align:center;display:block;color:#4b5563">R$ 280 <br/><span style="color:#9fa6b2;text-transform:uppercase;font-size:12px;margin-top:2px">R$ 247</span></span></td></tr></tbody></table></td></tr></tbody></table></div></div></div><div style="text-align:left"><p style="color:#333333;margin-bottom:3px"><b>Abraços, <br/>Equipe VOA Hotéis </b></p><p style="color:#333333;margin-bottom:3px">Atlântica 1130, 8º andar, CEP 22021-000 - <a style="color:#6366f1;margin-bottom:3px" href="https://www.voahoteis.com.br">https://www.voahoteis.com.br</a></p><img src="https://voa-hoteis-assets.s3.amazonaws.com/nps/images/satlite-white.png" alt="VOA Hotéis" width="130" style="text-align:center;display:block;margin:0 auto"/></div></div>`,
-  });
-
-  res.status(200).json({ message: "Test email sent" });
-}
